task 9, fix bugs, typo

+++ python-more_data_structures/9-multiply_by_2.py
@@ -3,5 +3,5 @@ def multiply_by_2(a_dictionary):
     pictionary = {}
     for k, v in a_dictionary.items():
         if v % 2 == 0:
-            pictionary[k] = int(v * 2)
+            pictionary[k] = v * 2
     return pictionary
